+++ main.py
@@ -1,24 +1,13 @@
 import tkinter as tk
 from tkinter import ttk
 
-import matplotlib.pyplot as plt
-
 import logging
 
 from tabs.tab3 import create_tab3
 from tabs.tab2 import create_tab2
 from tabs.tab1 import create_tab1
 
-
-def configure_matplotlib():
-    """Configure global matplotlib settings used across the application."""
-    plt.rcParams['font.family'] = 'Times New Roman'
-    plt.rcParams['font.size'] = 14  # Общий размер шрифта
-    plt.rcParams['axes.titlesize'] = 14  # Размер шрифта для заголовков осей
-    plt.rcParams['axes.labelsize'] = 14  # Размер шрифта для подписей осей
-    plt.rcParams['xtick.labelsize'] = 14  # Размер шрифта для меток оси X
-    plt.rcParams['ytick.labelsize'] = 14  # Размер шрифта для меток оси Y
-    plt.rcParams['legend.fontsize'] = 14  # Размер шрифта для легенды
+from settings import configure_matplotlib
 
 
 def on_closing(root):
